% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactionData.R
\docType{data}
\name{reactionData}
\alias{reactionData}
\title{Data of ruler drop test reaction time}
\format{
A tibble with 103 rows and 4 variables:
\describe{
  \item{student}{chr - Names of the students who carried out the test}
  \item{test}{dbl - test ID, each students carried out more than 3 tests}
  \item{time_before}{dbl - denoting average distance the meterstick fell in cm before drinking coke}
  \item{time_after}{dbl - denoting average distance the meterstick fell in cm after drinking coke}
}
}
\source{
\url{https://www.topendsports.com/testing/tests/reaction-stick.htm}
}
\usage{
reactionData
}
\description{
This test uses the known properties of gravity to determine how long it takes a person to respond to the dropping of an object by measuring how far the object can falls before being caught. This data was collected from studnts before and after drinking coke.
}
\keyword{datasets}
