% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{rgirls_colours}
\alias{rgirls_colours}
\title{R Girls colours}
\usage{
rgirls_colours(cols = NULL)
}
\arguments{
\item{cols}{Name(s) of colour(s) in snake case to extract from the R Girls colour palette (e.g. "dark_pink").
If `NULL`, the function returns all the colours in the palette/}
}
\value{
A named vector of colours
}
\description{
R Girls colours
}
\examples{
rgirls_colours(c("dark_pink", "goa_green"))
}
