% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGirls-package.R
\docType{package}
\name{RGirls-package}
\alias{RGirls}
\alias{RGirls-package}
\title{RGirls: RGirlsData}
\description{
This package contains datasets for teaching purposes from RGirls Community.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-girls.github.io/RGirls-pkg/}
}

}
\author{
\strong{Maintainer}: Batool Almarzouq \email{batool@liverpool.ac.uk} (\href{https://orcid.org/0000-0002-3905-2751}{ORCID})

}
\keyword{internal}
